+++ ExtendedTooltip/Systems/ExtendedTempTooltipSystem.cs
@@ -1,11 +1,16 @@
 ﻿using ExtendedTooltip.Settings;
 using Game.Common;
+using Game.Prefabs;
 using Game.Simulation;
 using Game.Tools;
 using Game.UI.Tooltip;
+using Gooee.Plugins;
+using HarmonyLib;
 using System;
 using System.Linq;
+using System.Reflection;
 using Unity.Entities;
+using Unity.Mathematics;
 using UnityEngine.Scripting;
 
 namespace ExtendedTooltip.Systems
@@ -18,12 +23,13 @@ namespace ExtendedTooltip.Systems
         private ExtendedTooltipSystem m_ExtendedTooltipSystem;
         private ToolRaycastSystem m_ToolRaycastSystem;
         private TerrainSystem m_TerrainSystem;
+        private TerrainToolSystem m_TerrainToolSystem;
 
         private LocalSettings m_Settings;
         private StringTooltip m_NetToolMode;
 
-        private System.Reflection.Assembly[] loadedAssemblies;
-        private bool waterFeaturesLoaded;
+        private Type _waterFeaturesType;
+        private Assembly[] loadedAssemblies;
 
         [Preserve]
         protected override void OnCreate()
@@ -34,12 +40,13 @@ namespace ExtendedTooltip.Systems
             m_NetToolSystem = World.GetOrCreateSystemManaged<NetToolSystem>();
             m_CustomTranslationSystem = World.GetOrCreateSystemManaged<CustomTranslationSystem>();
             m_ExtendedTooltipSystem = World.GetOrCreateSystemManaged<ExtendedTooltipSystem>();
+            m_TerrainToolSystem = World.GetOrCreateSystemManaged<TerrainToolSystem>();
             m_TerrainSystem = World.GetOrCreateSystemManaged<TerrainSystem>();
             m_ToolRaycastSystem = World.GetOrCreateSystemManaged<ToolRaycastSystem>();
 
             loadedAssemblies = AppDomain.CurrentDomain.GetAssemblies();
-            waterFeaturesLoaded = IsAssemblyLoaded("Water_Features");
             m_Settings = m_ExtendedTooltipSystem.m_LocalSettings;
+            DetectWaterFeaturesMod();
 
             m_NetToolMode = new StringTooltip
             {
@@ -51,7 +58,7 @@ namespace ExtendedTooltip.Systems
         protected override void OnUpdate()
         {
             ModSettings modSettings = m_Settings.ModSettings;
-            if (m_ToolSystem != null)
+            if (m_ToolSystem != null && m_ToolSystem.activeTool is not DefaultToolSystem)
             {
                 if (m_ToolSystem.activeTool is NetToolSystem && modSettings.ShowNetToolSystem && modSettings.ShowNetToolMode)
                 {
@@ -71,36 +78,51 @@ namespace ExtendedTooltip.Systems
 
                 if ((modSettings.ShowTerrainToolHeight && m_ToolSystem.activeTool is TerrainToolSystem) || (modSettings.ShowWaterToolHeight && (m_ToolSystem.activeTool is WaterToolSystem || AnyOtherSupportedCustomToolSystem())))
                 {
+                    TerrainHeightData heightData = m_TerrainSystem.GetHeightData();
+
                     if (m_ToolRaycastSystem.GetRaycastResult(out RaycastResult raycastResult))
                     {
-                        TerrainHeightData heightData = m_TerrainSystem.GetHeightData();
-                        int height = (int) TerrainUtils.SampleHeight(ref heightData, raycastResult.m_Hit.m_HitPosition);
+                        int applyHeight = (int) TerrainUtils.SampleHeight(ref heightData, raycastResult.m_Hit.m_HitPosition);
                         StringTooltip terrainToolMode = new()
                         {
                             icon = "Media/Glyphs/Slope.svg",
-                            path = "terrainTool",
-                            value = m_CustomTranslationSystem.GetLocalGameTranslation("Common.VALUE_METER", "-", "SIGN", "", "VALUE", height.ToString()),
+                            path = "terrainToolApplyHeight",
+                            value = m_CustomTranslationSystem.GetLocalGameTranslation("Common.VALUE_METER", "-", "SIGN", "", "VALUE", applyHeight.ToString()),
                         };
                         AddMouseTooltip(terrainToolMode);
+
+                        // Add level height tooltip if the tool is set to level
+                        if (m_ToolSystem.activeTool is TerrainToolSystem && m_TerrainToolSystem.prefab.m_Type == TerraformingType.Level)
+                        {
+                            float3 targetPosition = Traverse.Create(m_TerrainToolSystem).Field<float3>("m_TargetPosition").Value;
+                            int targetHeight = (int)TerrainUtils.SampleHeight(ref heightData, targetPosition);
+                            StringTooltip terrainTargetPosition = new()
+                            {
+                                icon = "Media/Tools/Terrain Tool/Level.svg",
+                                path = "terrainToolTargetHeight",
+                                color = targetHeight > applyHeight ? TooltipColor.Success : targetHeight == applyHeight ? TooltipColor.Info : TooltipColor.Error,
+                                value = m_CustomTranslationSystem.GetLocalGameTranslation("Common.VALUE_METER", "-", "SIGN", "", "VALUE", targetHeight.ToString()),
+                            };
+                            AddMouseTooltip(terrainTargetPosition);
+                        }
                     }
                 }
             }
         }
 
-        private bool IsAssemblyLoaded(string assemblyName)
+        private void DetectWaterFeaturesMod()
         {
-            return loadedAssemblies.Any(assembly => assembly.FullName.StartsWith(assemblyName));
+            _waterFeaturesType = loadedAssemblies.FirstOrDefault(a => a.GetName().Name == "Water_Features")?
+                .GetTypes()
+                .FirstOrDefault(t => t.FullName == "Water_Features.Tools.CustomWaterToolSystem");
         }
 
         private bool AnyOtherSupportedCustomToolSystem()
         {
-            // Support for WaterFeatures mod
-            /*if (waterFeaturesLoaded && m_ToolSystem.activeTool is Water_Features.Tools.CustomWaterToolSystem)
-            {
-                return true;
-            }*/
+            if (_waterFeaturesType == null)
+                return false;
 
-            return false;
+            return m_ToolSystem.activeTool.GetType() == _waterFeaturesType;
         }
 
         protected override void OnCreateForCompiler()